requirements.txt
mlb-statsapi
psycopg2-binary
redis
requests
ollama
